Rename times and prices arrays
The arrays for times and prices were named times_open, times_close,
prices_member and prices_nonmember. However, it is much more sensible for them
to be associative arrays and they are such in the processing of the data before
the SQL queries. Now they are associative arrays everywhere. Their first
dimension is the day the second is type - member/nonmember and open/close.