Parser fills database tables clubs and club_schedule
The parser now truncates and fills the database tables clubs and club_schedule.
The club_schedule represents time schedule and prices for club for given day.
Currently only the time information is entered, not the prices.
If the day is selected but no time is given, the opening and closing times are
set to null. If all the days are selected, no row is inserted - the information
is filled in the clubs table instead.

Change club array structure - now it is more like the structure of the XML where
there are no opening times and closing times - just a raw field with the times.
The same goes for the prices.

Rename the day and hour regular expression, appending _regex at the end. Before
it was easy to forget about them and reassign them.

Check whether there is another club with the same name before inserting to the
database. Now the successive clubs with the same name are skipped, maybe somehow
the previous record can be edited, although this can remove the old data and
would probably not be supported.