Fix bug in query.php when sports empty
If the sports array was empty, it did not bother to check if there are other
supplied arguments. Now all arguments are first-class citizens instead of being
discriminated against. The sports are just assumed to be there, and if they are
not, their default value as empty array is fine too. Only when it is the only
selected argument does it fall back to simple getClubsBySports.