Fix sports regex
Fix the sports regex changing the place of operands in disjoint of regex.
The problem was that the empty string matched the global regex so the days
regex could never be reached. Moving the days regex before the global regex
fixes this problem because the days must be present in the string to be
matched by the days regex.