Different names for different lines
Different names for the subpatterns, containing the line number in the end.
Before this change, the last line overwrote the previous. This can be done in
other means, like giving a name to the line and then parsing the lines one by
one. However, this one is probably easier to do for now. It should be changed
in the future.