Fix hour regex and modify time check
The hours regex was matching an additional whitespace when an hour is selected
without selecting minutes and am/pm/noon (i.e. in 24 mode without minutes).
Time check is changed a bit, now it handles point instead of colon. Colon was
matched by the hour regex before but was not handled in time check.
The testing code is changed to show the different possible hour formats.