parser, general: Log messages, print to console
Save messages in error log. Not all are errors.
For example, the message 'XMLs are the same, database not updated'
indicates that we do not need to parse the XMLs because they have
been parsed and did not change in the meantime.
If CLI (console) mode detected, print without formatting.
Otherwise, apply formatting.